pull helper functions into smaller files
