+++ cct/replica_test.go
@@ -1,119 +1,30 @@
+/*
+ Copyright 2017 Crunchy Data Solutions, Inc.
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+*/
+
+// Crunchy Container Test
 package cct
 
 import (
-	"database/sql"
-	"fmt"
+	// "fmt"
 	"testing"
-	// "time"
-    "github.com/docker/docker/client"
-
-    _ "github.com/lib/pq"
 )
 
-func isReplicationStarted(
-	docker *client.Client,
-	containerId string) (ok bool, err error) {
-
-    conStr, err := buildConnectionString(
-	    docker, containerId, "postgres", "postgres")
-    if err != nil {
-        return
-    }
-
-	pg, _ := sql.Open("postgres", conStr)
-    defer pg.Close()
-
-    query := `SELECT EXISTS (SELECT 1 from pg_stat_replication
-    	WHERE application_name = 'replica' and state = 'streaming');`
-
-    err = pg.QueryRow(query).Scan(&ok)
-    if err != nil {
-    	return
-    }
-
-    return
-}
-
-func replSentEqReplay(conStr string) (ok bool, err error) {
-
-	pg, _ := sql.Open("postgres", conStr)
-    defer pg.Close()
-
-    query := `SELECT (sent_location = replay_location) from pg_stat_replication 
-    	WHERE application_name='replica' and state='streaming';`
-
-    err = pg.QueryRow(query).Scan(&ok)
-    if err != nil {
-    	return
-    }
-
-    return
-}
-
-func waitForReplay(
-	docker *client.Client, 
-	containerId string,
-	timeoutSeconds int64) (err error) {
-
-    conStr, err := buildConnectionString(
-	    docker, containerId, "postgres", "postgres")
-    if err != nil {
-        return
-    }
-
-    fmt.Printf("Waiting maximum of %d seconds for replay", timeoutSeconds)
-
-    escape := func() (bool, error) {
-    	irs, err := isReplicationStarted(docker, containerId)
-    	return ! irs, err
-    }
-    condition1 := func() (bool, error) {
-    	return replSentEqReplay(conStr)
-    }
-    if ok, err := timeoutOrReady(
-    	timeoutSeconds,
-        escape,
-        []func() (bool, error){condition1},
-        500); err != nil {
-        return err
-    } else if ! ok {
-        return fmt.Errorf("replication stopped; or timeout expired, and replay has not completed.")
-    }
-
-    return
-}
-
-func startMasterReplica(
-    t *testing.T,
-    docker *client.Client,
-    basePath string,
-    timeoutSeconds int64) (masterId, replicaId string, cleanup func(skip bool)) {
-
-    fmt.Println("Starting master-replica example, and pausing while example sleeps for 20 seconds")
-    cleanup = startDockerExampleForTest(t, basePath, "master-replica")
-
-    var err error
-
-    fmt.Printf("\nWaiting maximum of %d seconds for master container", timeoutSeconds)
-    masterId, err = waitForPostgresContainer(docker, "master", timeoutSeconds)
-    if err != nil {
-        t.Fatal("master container did not start")
-    }
-    t.Log("Started master container: " + masterId)
-
-    fmt.Printf("\nWaiting maximum of %d seconds for replica container", timeoutSeconds)
-    replicaId, err = waitForPostgresContainer(docker, "replica", timeoutSeconds)
-    if err != nil {
-        t.Fatal("replica container did not start")
-    }
-    t.Log("Started replica container: " + replicaId)
-
-    return
-}
 
 func TestDockerReplica(t *testing.T) {
 
-    const exampleName = "master-replica"
+    // const exampleName = "master-replica"
     const timeoutSeconds = 60
     const skipCleanup = false
 
@@ -125,8 +36,10 @@ func TestDockerReplica(t *testing.T) {
     masterId, replicaId, cleanup := startMasterReplica(t, docker, buildBase, timeoutSeconds)
     defer cleanup(skipCleanup)
 
+    conStr := conStrTestPostgres(t, docker, masterId)
+
     if t.Run("ReplicationStarted", func (t *testing.T) {
-    	if ok, err := isReplicationStarted(docker, masterId); err != nil {
+    	if ok, err := isReplicationStarted(conStr); err != nil {
     		t.Error(err)
     	} else if ! ok {
     		t.Error("Replication has not started")
@@ -136,24 +49,24 @@ func TestDockerReplica(t *testing.T) {
     }
 
     t.Log("Write some data to master")
-    facts, err := writeSomeData(docker, masterId)
+    facts, err := writeSomeData(docker, masterId, userdb)
     if err != nil {
         t.Error(err)
     }
-    t.Log(facts)
 
     t.Log("Waiting for replay")
-    if err := waitForReplay(docker, "master", timeoutSeconds); err != nil {
+    if err := waitForReplay(conStr, timeoutSeconds); err != nil {
     	t.Fatal(err)
     }
 
     t.Run("CheckReplica", func (t *testing.T) {
-        ok, foundrc, err := assertSomeData(docker, replicaId, facts.rowcount)
+        ok, found, err := assertSomeData(docker, replicaId, userdb, facts)
         if err != nil {
         	t.Error(err)
         }
         if ! ok {
-        	t.Errorf("Expected %d rows; found %d\n", facts.rowcount, foundrc)
+        	t.Errorf("Expected %n rows, %n bytes\nfound %n rows, %n bytes\n",
+                facts.rowcount, facts.relsize, found.rowcount, found.relsize)
         }
     })
